♻️refactor: Summary of file types related to features/api detail page - 상세페이지 관련 api 타입 부분 정리 Related to: #126
